Took out most of the crap
Made the site tons more light weight
