--- .istanbul.yml
@@ -1,7 +1,0 @@
-instrumentation:
-  root: src
-  include-all-sources: true
-  verbose: true
-  excludes: ["./src/**/*.test.js" ,"./build"]
-reporting:
-  dir: "coverage"
