+++ package.json
@@ -1,74 +1,15 @@
 {
   "name": "menas.co",
-  "productName": "menas.co - A Journaling App",
-  "version": "0.0.1",
-  "description": "A Simple Native/Mobile Journaling App",
-  "license": "MIT",
-  "repository": "iammenasco/menas.co",
-  "author": {
-    "name": "Brian Menasco",
-    "email": "menasco@me.com",
-    "url": "http://menas.co"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "electronVersion": "0.34.0",
+  "version": "1.0.0",
+  "description": "Landing Page for menas.co",
+  "main": "webpack.config.js",
   "scripts": {
-    "test": "snyk protect && snyk test && xo; babel-node ./node_modules/istanbul/lib/cli.js --include-all-sources cover mocha -- ./src/**/*.test.js -R spec --recursive --report lcov",
-    "update": "ncu --upgradeAll",
-    "build": "export NODE_ENV='production'; electron-packager . $npm_package_productName --out=dist --ignore='^/dist$' --prune --asar --version=$npm_package_electronVersion",
-    "build-sass": "node-sass --include-path src/styles --source-map index.map src/styles/index.scss index.css",
-    "watch-sass": "nodemon -e scss -x \"npm run build-sass\"",
-    "start": "export NODE_ENV='dev'; electron . & webpack --watch --progress --colors --output-source-map-file & npm run watch-sass",
-    "snyk-protect": "snyk protect",
-    "prepublish": "npm run snyk-protect"
+    "start": "http-server -p ${PORT:=3000}"
   },
-  "files": [
-    "index.js",
-    "index.html",
-    "index.css"
-  ],
-  "keywords": [
-    "electron-app"
-  ],
-  "dependencies": {
-    "babel-cli": "^6.22.2",
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
-    "browser-sync": "^2.18.7",
-    "browser-sync-webpack-plugin": "^1.1.4",
-    "codeclimate-test-reporter": "^0.4.0",
-    "css-loader": "^0.26.1",
-    "electron": "^1.4.15",
-    "electron-debug": "^1.1.0",
-    "electron-packager": "^8.5.1",
-    "expect.js": "^0.3.1",
-    "iammenasco-blue": "^1.0.0",
-    "istanbul": "^1.1.0-alpha.1",
-    "keen-tracking": "^1.1.3",
-    "mocha": "^4.0.0",
-    "node-sass": "^4.5.0",
-    "nodemon": "^1.11.0",
-    "opbeat": "^4.11.0",
-    "sass-loader": "^4.1.1",
-    "snyk": "^1.62.2",
-    "style-loader": "^0.13.1",
-    "tap": "^10.0.2",
-    "webpack": "1.14.0",
-    "xo": "^0.17.1"
-  },
-  "xo": {
-    "ignores": [
-      "build/*",
-      "**/*.test.js"
-    ],
-    "esnext": true,
-    "envs": [
-      "node",
-      "browser"
-    ],
-    "space": 2
+  "license": "MIT",
+  "devDependencies": {
+    "http-server": "^0.10.0",
+    "snyk": "^1.62.2"
   },
   "snyk": true
 }
