--- src/styles/auroral.scss
@@ -1,4 +1,0 @@
-@import 'config';
-@import 'animations';
-@import 'backgrounds';
-@import 'stars';
